Refactor fabric classes
