+++ fabric/src/main/java/io/github/retrooper/packetevents/factory/fabric/FabricPacketEventsAPI.java
@@ -21,7 +21,6 @@ package io.github.retrooper.packetevents.factory.fabric;
 import com.github.retrooper.packetevents.PacketEvents;
 import com.github.retrooper.packetevents.PacketEventsAPI;
 import com.github.retrooper.packetevents.injector.ChannelInjector;
-import com.github.retrooper.packetevents.manager.InternalPacketListener;
 import com.github.retrooper.packetevents.manager.player.PlayerManager;
 import com.github.retrooper.packetevents.manager.protocol.ProtocolManager;
 import com.github.retrooper.packetevents.manager.registry.RegistryManager;
@@ -34,6 +33,9 @@ import com.github.retrooper.packetevents.util.LogManager;
 import io.github.retrooper.packetevents.LazyHolder;
 import io.github.retrooper.packetevents.impl.netty.NettyManagerImpl;
 import io.github.retrooper.packetevents.impl.netty.manager.player.PlayerManagerAbstract;
+import io.github.retrooper.packetevents.manager.FabricLogger;
+import io.github.retrooper.packetevents.manager.FabricProtocolManager;
+import io.github.retrooper.packetevents.manager.FabricServerManager;
 import io.github.retrooper.packetevents.manager.InternalFabricPacketListener;
 import io.github.retrooper.packetevents.manager.registry.FabricRegistryManager;
 import net.fabricmc.api.EnvType;
