+++ fabric/src/main/java/io/github/retrooper/packetevents/manager/FabricLogger.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.github.retrooper.packetevents.factory.fabric;
+package io.github.retrooper.packetevents.manager;
 
 import com.github.retrooper.packetevents.util.LogManager;
 import net.kyori.adventure.text.format.NamedTextColor;
